requirements.txt
qiskit==2.2.2
qiskit-machine-learning==0.9.0
qiskit-aer==0.17.2
qiskit-algorithms==0.4.0
streamlit==1.54.0
pandas==2.2.3
scikit-learn==1.7.2
numpy==2.1.3
plotly==6.5.2
pytest==9.0.2
fastapi==0.134.0
uvicorn==0.41.0
pydantic==2.10.6

joblib
hvac
pyodbc
azure-core
confluent-kafka
opentelemetry-instrumentation-openai